+++ src/causaltensor/cauest/DebiasConvex.py
@@ -191,7 +191,7 @@ class DCPanelSolver(PanelSolver):
                 if (np.linalg.matrix_rank(M) > suggest_r):
                     M_debias, tau_debias = self.debias(pre_M, pre_tau, l*coef)
                     if method_non_neg:
-                        M = util.SVD_non_negative(M, suggest_r)
+                        M = util.SVD_non_negative(M_debias, suggest_r)
                     else:
                         M = util.SVD(M_debias, suggest_r)
                     tau = tau_debias 
@@ -203,6 +203,9 @@ class DCPanelSolver(PanelSolver):
             M, tau = self.non_convex_PR(suggest_r, initial_tau = pre_tau, method_non_neg=method_non_neg)
 
         if method == 'auto':
+            '''
+                select the best estimator from non-convex and convex
+            '''
             M1, tau1 = self.non_convex_PR(suggest_r, initial_tau = self.solve_tau(self.O), method_non_neg=method_non_neg)
             if np.linalg.matrix_rank(M) != suggest_r or np.linalg.norm(self.O-M-np.tensordot(self.Z, tau,  axes=([2], [0]))) > np.linalg.norm(self.O-M1-np.tensordot(self.Z, tau1,  axes=([2], [0]))):
                 M = M1
@@ -216,10 +219,10 @@ class DCPanelSolver(PanelSolver):
             return M, tau, standard_deviation
 
 
-    def DC_PR_auto_rank(self, spectrum_cut = 0.002, method='convex'):
+    def DC_PR_auto_rank(self, spectrum_cut = 0.002, method='convex', method_non_neg=None):
         s = np.linalg.svd(self.O, full_matrices = False, compute_uv=False)
         suggest_r = np.sum(np.cumsum(s**2) / np.sum(s**2) <= 1-spectrum_cut)
-        return self.DC_PR_with_suggested_rank(suggest_r = suggest_r, method=method)
+        return self.DC_PR_with_suggested_rank(suggest_r = suggest_r, method=method, method_non_neg=method_non_neg)